update clientapi test to use service cient API

+++ test/core/clientapi/test_client.c
@@ -124,7 +124,7 @@ int write_to_socket(char *address, char* port, char* filename, char* endpoint)
    strcat(buffer_l, port);
    strcat(buffer_l, "\r\n");
    strcat(buffer_l, "Content-Length: ");
-   sprintf(tmpstr, "%d", strlen(buffer));
+   sprintf(tmpstr, "%d", (int)strlen(buffer));
    strcat(buffer_l, tmpstr);
    strcat(buffer_l, "\r\n");
    /*strcat(buffer_l, "SOAPAction: http://localhost:9090/axis2/services/echo/echo\r\n");*/
